worker/requirements.txt
strawberry-graphql
skema